+++ src/service/payment/movieTicket.ts
@@ -487,7 +487,7 @@ export function refundMovieTicket(params: factory.task.IData<factory.taskName.Re
 
         // 本アクションに対応するPayActionを取り出す
         const payAction = await findPayActionByOrderNumber<factory.paymentMethodType.MGTicket | factory.paymentMethodType.MovieTicket>({
-            object: { typeOf: paymentMethodType, paymentMethodId: params.object.paymentMethodId },
+            object: { paymentMethod: paymentMethodType, paymentMethodId: params.object.paymentMethodId },
             purpose: { orderNumber: params.purpose.orderNumber }
         })(repos);