+++ src/service/payment/paymentCard.ts
@@ -318,7 +318,7 @@ export function voidTransaction(
                     .then((actions) => actions
                         // tslint:disable-next-line:no-suspicious-comment
                         // TODO Chevre決済カードサービスに対して動的にコントロール
-                        .filter((a) => a.object.typeOf === factory.paymentMethodType.PaymentCard)
+                        .filter((a) => a.object.paymentMethod === factory.paymentMethodType.PaymentCard)
                     );
         }
 
@@ -399,7 +399,7 @@ export function refundPaymentCard(params: factory.task.IData<factory.taskName.Re
     }) => {
         // 本アクションに対応するPayActionを取り出す
         const payAction = await findPayActionByOrderNumber<typeof params.object.typeOf>({
-            object: { typeOf: params.object.typeOf, paymentMethodId: params.object.paymentMethodId },
+            object: { paymentMethod: params.object.typeOf, paymentMethodId: params.object.paymentMethodId },
             purpose: { orderNumber: params.purpose.orderNumber }
         })(repos);